backend-improved

+++ src/app/pages/employees/employees.component.ts
@@ -5,6 +5,7 @@ import { baseUrl } from 'src/app/environments/environment';
 import { GetEmployee } from 'src/app/models/getEmployee';
 import { DatePipe } from '@angular/common';
 import { ToastrService } from 'ngx-toastr';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-employees',
@@ -25,17 +26,29 @@ export class EmployeesComponent implements OnInit {
     this.route.navigate(['/addemployee']);
   }
   onDeleteEmployee(id: string) {
-    this.http.delete(`${baseUrl}Employee/delete/${id}`).subscribe(
-      () => {
-        console.log('Employee deleted successfully');
-        this.getEmployee();
-        this.toastr.success('Employee Deleted', 'Operation Successful');
-        this.route.navigate(['/employees']);
-      },
-      (error) => {
-        console.error('Employee deleting food:', error);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You are about to delete this Employee from list. This action cannot be undone',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.http.delete(`${baseUrl}Employee/delete/${id}`).subscribe(
+          () => {
+            console.log('Employee deleted successfully');
+            this.getEmployee();
+            this.toastr.success('Employee Deleted', 'Operation Successful');
+            this.route.navigate(['/employees']);
+          },
+          (error) => {
+            console.error('Employee deleting food:', error);
+          }
+        );
       }
-    );
+    });
   }
 
   getEmployee() {
